requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
click==7.1.2
cycler==0.10.0
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
pathspec==0.8.1
pigpio==1.78
Pillow==8.1.0
pylint==2.6.0
pyparsing==2.4.7
PySimpleGUI==4.34.0
python-dateutil==2.8.1
pyzmq==21.0.2
regex==2020.11.13
scipy==1.6.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
